修改页面生成的方式，由异步改为同步

+++ src/main/java/com/scsme/dataConfigCenter/util/HTMLTemplateUtil.java
@@ -23,48 +23,46 @@ import java.util.Map;
 public class HTMLTemplateUtil {
     private static final String TEMPLATES_DIR = "/templates/";
     private static final String FREEMARKER_DIR = "/templates/freemarker/";
+    private static final String PREVIEW_TEMPLATES_DIR = "/templates/preview/";
+    private static final String PREVIEW_FREEMARKER_DIR = "/templates/preview/freemarker/";
     private static final String TEMPLATE_SUFFIX = ".ftl";
     private static final String HTML_SUFFIX = ".html";
     private static final String TITLE = "title";
     private static final String COMPONENTS = "components";
 
-    public static void deleteHTMLFile(Long layoutId, String url, LayoutMapper layoutMapper) {
-        try {
-            File file = new File(new ClassPathResource(TEMPLATES_DIR).getURI().getPath() + url + HTML_SUFFIX);
-            file.delete();
-            if (layoutMapper != null) {
-                layoutMapper.removeLayoutFile(layoutId);
-            }
-        } catch (Exception e) {
-            log.error("删除HTML失败....");
-            log.error(e.getLocalizedMessage());
-            e.printStackTrace();
+    public static void deleteHTMLFile(Long layoutId, String url, LayoutMapper layoutMapper, Boolean isPreview) throws Exception {
+        File file = new File(new ClassPathResource(isPreview ? TEMPLATES_DIR : PREVIEW_TEMPLATES_DIR).getURI().getPath() + url + HTML_SUFFIX);
+        file.delete();
+        if (layoutMapper != null && !isPreview) {
+            layoutMapper.removeLayoutFile(layoutId);
         }
     }
 
-    public static void generateHTMLFile(LayoutVO layout, LayoutMapper layoutMapper) {
+    public static void deleteHTMLFile(Long layoutId, String url, LayoutMapper layoutMapper) throws Exception {
+        deleteHTMLFile(layoutId, url, layoutMapper, false);
+    }
+
+    public static void generateHTMLFile(LayoutVO layout, LayoutMapper layoutMapper, Boolean isPreview) throws Exception {
         //写入到本地文件
         List<ComponentVO> components = layout.getComponents();
         Configuration configuration = new Configuration(Configuration.VERSION_2_3_30);
-        try {
-            configuration.setDirectoryForTemplateLoading(new File(new ClassPathResource(FREEMARKER_DIR).getURI().getPath()));
-            Template template = configuration.getTemplate(layout.getTemplateName() + TEMPLATE_SUFFIX);
-            Map<String, Object> dataMap = new HashMap<>();
-            dataMap.put(TITLE, layout.getTitle());
-            dataMap.put(COMPONENTS, components);
-            File file = new File(new ClassPathResource(TEMPLATES_DIR).getURI().getPath() + layout.getUrl() + HTML_SUFFIX);
-            file.delete();
-            try (Writer out = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file)))) {
-                template.process(dataMap, out);
-            }
-            if (layoutMapper != null) {
-                layout.setFile(Files.readAllBytes(Paths.get(file.getPath())));
-                layoutMapper.updateById(layout);
-            }
-        } catch (Exception e) {
-            log.error("生成HTML失败....");
-            log.error(e.getLocalizedMessage());
-            e.printStackTrace();
+        configuration.setDirectoryForTemplateLoading(new File(new ClassPathResource(isPreview ? FREEMARKER_DIR : PREVIEW_FREEMARKER_DIR).getURI().getPath()));
+        Template template = configuration.getTemplate(layout.getTemplateName() + TEMPLATE_SUFFIX);
+        Map<String, Object> dataMap = new HashMap<>();
+        dataMap.put(TITLE, layout.getTitle());
+        dataMap.put(COMPONENTS, components);
+        File file = new File(new ClassPathResource(isPreview ? TEMPLATES_DIR : PREVIEW_TEMPLATES_DIR).getURI().getPath() + layout.getUrl() + HTML_SUFFIX);
+        file.delete();
+        try (Writer out = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(file)))) {
+            template.process(dataMap, out);
         }
+        if (layoutMapper != null && !isPreview) {
+            layout.setFile(Files.readAllBytes(Paths.get(file.getPath())));
+            layoutMapper.updateById(layout);
+        }
+    }
+
+    public static void generateHTMLFile(LayoutVO layout, LayoutMapper layoutMapper) throws Exception {
+        generateHTMLFile(layout, layoutMapper, false);
     }
 }
